Store datetimes as GMT and display in local timezone
Also nicified logging of proposal time changes and enabled setting
time to a one (1) minute granularity. Known issue: You need to
fiddle with the controls to select the same hour as previously
selected in the datetime picker. Worthy of note that I added mrt:
moment-timezone here, but we're not yet actually using any features
which weren't already in moment (which is still here but unused).