Automatically run migrations on server startup
We need to do it this way for them to work anyway because Meteor's
Collection.update will only update one record at a time when invoked
as untrusted code (e.g. from the client console). Since we want to
update multiple records, we run it on the server, and we also gain
the benefit of not having to go manually run it. Magic!