Migrations: Finally working nicely
The issues were firstly that the code was in the wrong place. Placing
it in Meteor.startup in client/main.js caused it to execute on every
page load. Secondly, the Collection.update method only updates a single
record if the code is untrusted (e.g. the client console), but in any
case to update multiple records, we need to pass the {multi: true} flag.
Finally, moving the migration method definitions to server/main.js keeps
them all in one place by convention (we may eventually give them their
own file).