Preventing empty comments...seems pretty solid now
Because the previous commit made use of $pristine, it only worked
on the first comment each page load. Subsequent comments were not
pristine and had not been validated and were therefore possible to
submit empty. I guess we could validate before submit, or disallow
empty comments with a schema, but what I did here instead was use
the same setPristineAndUntouched function we're using when the main
proposal form is submitted. Since the logic is identical we'll want
to DRY this out eventually.